requirements.txt
sqlalchemy==1.4.25
pyyaml==5.3.1
numpy==1.19.5
pandas==1.3.1
yfinance==0.1.63
ta==0.7.0